Be more lenient when stripping frameworks and dSYMs for non fat binaries

+++ lib/cocoapods/generator/embed_frameworks_script.rb
@@ -43,10 +43,14 @@ module Pod
 
           SWIFT_STDLIB_PATH="${DT_TOOLCHAIN_DIR}/usr/lib/swift/${PLATFORM_NAME}"
 
+          # Used as a return value for each invocation of `strip_invalid_archs` function.
+          STRIP_BINARY_RETVAL=0
+
           # This protects against multiple targets copying the same framework dependency at the same time. The solution
           # was originally proposed here: https://lists.samba.org/archive/rsync/2008-February/020158.html
           RSYNC_PROTECT_TMP_FILES=(--filter "P .*.??????")
 
+          # Copies and strips a vendored framework
           install_framework()
           {
             if [ -r "${BUILT_PRODUCTS_DIR}/$1" ]; then
@@ -95,7 +99,7 @@ module Pod
             fi
           }
 
-          # Copies the dSYM of a vendored framework
+          # Copies and strips a vendored dSYM
           install_dsym() {
             local source="$1"
             if [ -r "$source" ]; then
@@ -109,12 +113,17 @@ module Pod
 
               # Strip invalid architectures so "fat" simulator / device frameworks work on device
               if [[ "$(file "$binary")" == *"Mach-O dSYM companion"* ]]; then
-               strip_invalid_archs "$binary"
+                strip_invalid_archs "$binary"
               fi
 
-              # Move the stripped file into its final destination.
-              echo "rsync --delete -av "${RSYNC_PROTECT_TMP_FILES[@]}" --filter \\"- CVS/\\" --filter \\"- .svn/\\" --filter \\"- .git/\\" --filter \\"- .hg/\\" --filter \\"- Headers\\" --filter \\"- PrivateHeaders\\" --filter \\"- Modules\\" \\"${DERIVED_FILES_DIR}/${basename}.framework.dSYM\\" \\"${DWARF_DSYM_FOLDER_PATH}\\""
-              rsync --delete -av "${RSYNC_PROTECT_TMP_FILES[@]}" --filter "- CVS/" --filter "- .svn/" --filter "- .git/" --filter "- .hg/" --filter "- Headers" --filter "- PrivateHeaders" --filter "- Modules" "${DERIVED_FILES_DIR}/${basename}.framework.dSYM" "${DWARF_DSYM_FOLDER_PATH}"
+              if [[ $STRIP_BINARY_RETVAL == 1 ]]; then
+                # Move the stripped file into its final destination.
+                echo "rsync --delete -av "${RSYNC_PROTECT_TMP_FILES[@]}" --filter \\"- CVS/\\" --filter \\"- .svn/\\" --filter \\"- .git/\\" --filter \\"- .hg/\\" --filter \\"- Headers\\" --filter \\"- PrivateHeaders\\" --filter \\"- Modules\\" \\"${DERIVED_FILES_DIR}/${basename}.framework.dSYM\\" \\"${DWARF_DSYM_FOLDER_PATH}\\""
+                rsync --delete -av "${RSYNC_PROTECT_TMP_FILES[@]}" --filter "- CVS/" --filter "- .svn/" --filter "- .git/" --filter "- .hg/" --filter "- Headers" --filter "- PrivateHeaders" --filter "- Modules" "${DERIVED_FILES_DIR}/${basename}.framework.dSYM" "${DWARF_DSYM_FOLDER_PATH}"
+              else
+                # The dSYM was not stripped at all, in this case touch a fake folder so the input/output paths from Xcode do not reexecute this script because the file is missing.
+                touch "${DWARF_DSYM_FOLDER_PATH}/${basename}.framework.dSYM"
+              fi
             fi
           }
 
@@ -136,10 +145,18 @@ module Pod
           # Strip invalid architectures
           strip_invalid_archs() {
             binary="$1"
-            # Get architectures for current file
-            archs="$(lipo -info "$binary" | rev | cut -d ':' -f1 | rev)"
+            # Get architectures for current target binary
+            binary_archs="$(lipo -info "$binary" | rev | cut -d ':' -f1 | awk '{$1=$1;print}' | rev)"
+            # Intersect them with the architectures we are building for
+            intersected_archs="$(echo ${ARCHS[@]} ${binary_archs[@]} | tr ' ' '\\n' | sort | uniq -d)"
+            # If there are no archs supported by this binary then warn the user
+            if [[ -z "$intersected_archs" ]]; then
+              echo "warning: [CP] Vendored binary '$binary' contains architectures ($binary_archs) none of which match the current build architectures ($ARCHS)."
+              STRIP_BINARY_RETVAL=0
+              return
+            fi
             stripped=""
-            for arch in $archs; do
+            for arch in $binary_archs; do
               if ! [[ "${ARCHS}" == *"$arch"* ]]; then
                 # Strip non-valid architectures in-place
                 lipo -remove "$arch" -output "$binary" "$binary" || exit 1
@@ -149,6 +166,7 @@ module Pod
             if [[ "$stripped" ]]; then
               echo "Stripped $binary of architectures:$stripped"
             fi
+            STRIP_BINARY_RETVAL=1
           }
 
         SH
